The metadata info for columns drops the paths.
Having the complete paths in the dirs metadata prevented a safe
operation when moving the tables to different directories,
or just when loading tables from different places that they where
created.

One issue that appeared while doing the fix is that, when a new
column in-memory is inserted, it is not clear which should be the
default action (copy on-disk or just use the reference?).  Probably
we would need to setup a configuration file for specifying default
actions.  For the time being, the reference will be used.