removing comment names for some tests
As we reuse test methods for different contexts, using a base class to change behavior, the comment names make difficult to see the actual test that is failing (as the comment will remain the same for all contexts). By default nose prints the method name + class that unambiguously describes the test.